P13 Area Of Triangle code

+++ Challenges/challenges/src/P13SimpleInterest.java
@@ -6,11 +6,12 @@ public class P13SimpleInterest {
         System.out.println("Simple Interest Program ");
         System.out.print("Enter Principle : ");
         int P = input.nextInt();
-        System.out.print("Enter Time : ");int T = input.nextInt();
+        System.out.print("Enter Time : ");
+        int T = input.nextInt();
         System.out.print("Enter Rate : ");
         int R = input.nextInt();
 
-        float simpleInterest = (P*T*R)/100;
+        int simpleInterest = (P*T*R)/100;
         System.out.println("Simple Interest is : " + simpleInterest);
     }
 }
